pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solarnews"
version = "1.0.0"
description = "Global solar power news aggregator - 全球太阳能行业新闻聚合器"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Solar News Team", email = "solar@example.com"}
]
maintainers = [
    {name = "Solar News Team", email = "solar@example.com"}
]
keywords = [
    "solar",
    "photovoltaic", 
    "news",
    "aggregator",
    "energy",
    "renewable",
    "crawler",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "feedparser>=6.0.0",
    
    # Async HTTP client
    "aiohttp>=3.9.0",
    
    # Web scraping and dynamic content
    "playwright>=1.40.0",
    
    # Natural language processing
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "textblob>=0.17.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    
    # Environment management
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
redis = [
    "redis>=5.0.0",
]
all = [
    "solarnews[dev,postgres,redis]",
]

[project.scripts]
solarnews = "solarnews.cli:main"

[project.urls]
Homepage = "https://github.com/ankrwu/newsup-solar"
Documentation = "https://github.com/ankrwu/newsup-solar#readme"
Repository = "https://github.com/ankrwu/newsup-solar.git"
Issues = "https://github.com/ankrwu/newsup-solar/issues"

[tool.setuptools]
packages = ["solarnews", "src"]

[tool.setuptools.package-data]
"*" = ["*.py", "*.txt", "*.md"]
